Refactored

+++ srcs/ft_putsrc/ft_putchar_fd.c
@@ -6,17 +6,17 @@
 /*   By: mdeville <mdeville@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/28 16:13:15 by mdeville          #+#    #+#             */
-/*   Updated: 2017/12/04 23:23:44 by mdeville         ###   ########.fr       */
+/*   Updated: 2017/12/06 15:16:14 by mdeville         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-size_t	ft_putchar_fd(const int fd, t_token token, va_list ap)
+int		ft_putchar_fd(const int fd, t_token token, va_list ap)
 {
 	int		width;
 	char	c;
-	size_t	cpt;
+	int		cpt;
 
 	width = token.width;
 	c = va_arg(ap, int);
